Fixed code generation from thrift files (for CLI and VS), fixed other issues
