+++ lib/netcore/Tests/Thrift.Tests/Thrift.Tests.csproj
@@ -1,23 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <PreBuildEvent>
-    </PreBuildEvent>
-    <OutputType>Library</OutputType>
-    <LangVersion>7</LangVersion>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="CompareNETObjects" Version="3.9.0" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk">
-      <Version>15.5.0-preview-20170810-02</Version>
-    </PackageReference>
-    <PackageReference Include="MSTest.TestAdapter">
-      <Version>1.2.0-beta3</Version>
-    </PackageReference>
-    <PackageReference Include="MSTest.TestFramework">
-      <Version>1.2.0-beta3</Version>
-    </PackageReference>
-    <PackageReference Include="NSubstitute" Version="2.0.3" />
+    <PackageReference Include="CompareNETObjects" Version="4.3.0" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0" />
+    <PackageReference Include="MSTest.TestAdapter" Version="1.2.0" />
+    <PackageReference Include="MSTest.TestFramework" Version="1.2.0" />
+    <PackageReference Include="NSubstitute" Version="3.1.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\Thrift\Thrift.csproj" />
