+++ test/netcore/ThriftTest/.gitignore
@@ -1,2 +1,2 @@
 # ignore for autogenerated files
-/ThriftAsync
+/ThriftTest
